更改目录结构，更新了部分函数方法，界面已差不多完工

+++ helper/ClearHelper.vb
@@ -1,6 +1,9 @@
 ﻿
 Public Class ClearHelper
-    Public Shared Sub ClearMemory()
+    <DllImport("kernel32.dll", EntryPoint:="SetProcessWorkingSetSize")>
+    Public Shared Function SetProcessWorkingSetSize(process As IntPtr, minSize As Integer, maxSize As Integer) As Integer
+    End Function
+    Public Shared Sub ClearMemory(virtualRAMbool As Boolean)
         On Error Resume Next
         Task.Factory.StartNew(New Action(Sub()
 
@@ -11,6 +14,9 @@ Public Class ClearHelper
                                                  Try
                                                      If oprocess IsNot Process.GetCurrentProcess Then
                                                          If oprocess.Handle <> IntPtr.Zero Then
+                                                             If virtualRAMbool Then
+                                                                 SetProcessWorkingSetSize(oprocess.Handle, -1, -1)
+                                                             End If
                                                              Psapi.EmptyWorkingSet(New Kernel32.SafeObjectHandle(oprocess.Handle))
                                                          End If
                                                      End If
@@ -36,9 +42,22 @@ Public Class ClearHelper
         counters.NextValue()
         Return counters
     End Function
+    Public Shared Function GetAllSubKeyAddin(RootReg As RegistryKey, dirpath As String) As Dictionary(Of String, String)
+        On Error Resume Next
+        Dim new_dic As New Dictionary(Of String, String)
+        For Each strs As String In RootReg.GetSubKeyNames
+            new_dic.Add(dirpath, strs + "\" + dirpath)
+        Next
+        Return new_dic
+    End Function
+    Public Shared Sub SetReg(RootReg As RegistryKey, dirpath As String, keyname As String, keyvalue As Object, valuekind As RegistryValueKind)
+        On Error Resume Next
+        RootReg.OpenSubKey(dirpath, True).SetValue(keyname, keyvalue, valuekind)
+    End Sub
 
     Public Shared Function EnableGPU(open As Boolean) As Boolean
         Try
+            Dim x = Registry.Users
             Dim key As RegistryKey = Registry.LocalMachine.OpenSubKey("SYSTEM\CurrentControlSet\Services\Vxd", True)
             key.CreateSubKey("BIOS", True)
             key.OpenSubKey("BIOS", True)
@@ -64,6 +83,7 @@ Public Class ClearHelper
         End If
     End Sub
     Public Shared Sub RestartProcess(name As String, Optional autoRun As Boolean = True)
+        On Error Resume Next
         Dim ProcessPath As String
         Dim process_all As Process() = Process.GetProcesses()
         For Each every_proc As Process In process_all
@@ -76,14 +96,51 @@ Public Class ClearHelper
             End If
         Next
     End Sub
+    Public Shared Function SetHibernateEnabled(bool As Boolean) As Boolean
+        Try
+            Dim key As RegistryKey = Registry.LocalMachine.OpenSubKey("SYSTEM\CurrentControlSet\Control\Power", True)
+            key.SetValue("HibernateEnabled", If(bool, 1, 0), RegistryValueKind.DWord)
+
+            Return True
+        Catch ex As Exception
+            Return False
+        End Try
+    End Function
+    Public Shared Function SetHiberbootEnabled(bool As Boolean) As Boolean
+        Try
+            Dim key As RegistryKey = Registry.LocalMachine.OpenSubKey("SYSTEM\CurrentControlSet\Control\Session Manager\Power", True)
+            key.SetValue("HiberbootEnabled", If(bool, 1, 0), RegistryValueKind.DWord)
+
+            Return True
+        Catch ex As Exception
+            Return False
+        End Try
+    End Function
+    Public Shared Sub CLearSystemDir()
+        On Error Resume Next
+        ClearFiles(ClearConst.RECYCLE_BIN_FILE)
+        ClearFiles(ClearConst.WINDOWS_BT_FILE)
+        ClearFiles(ClearConst.WINDOWS_WS_FILE)
+    End Sub
     Public Shared Sub CLearAllNormalDir()
         On Error Resume Next
-        ClearSystemUpdateTempFiles(ClearConst.APPDATA_LOCAL_TEMP)
+        ClearFiles(ClearConst.APPDATA_LOCAL_TEMP)
+        ClearFiles(ClearConst.APPDATA_ROAMING_TEMP)
+        ClearFiles(ClearConst.CURRENT_TEMP)
+        ClearFiles(ClearConst.SYSTEM_TEMP)
+        ClearFiles(ClearConst.SYSTEM_TMP)
+        ClearFiles(ClearConst.SYSTEM_TMPDIR)
+        ClearFiles(ClearConst.USER_TEMP)
+        ClearFiles(ClearConst.USER_TMP)
+        ClearFiles(ClearConst.PREFETCH_FILE)
+        ClearFiles(ClearConst.INTERNET_CACHE_FILE)
     End Sub
-    Public Shared Sub ClearSystemUpdateTempFiles(dir As String)
+    Public Shared Sub ClearFiles(dir As String)
         On Error Resume Next
         Task.Factory.StartNew(New Action(Sub()
-                                             DeleteDir(dir)
+                                             If Directory.Exists(dir) Then
+                                                 DeleteDir(dir)
+                                             End If
                                          End Sub))
     End Sub
